Move go module to echo directory.
